filtering out current tenant in duplicate page to drawer + using hook for tenant lookups
