+++ src/fields/tenantField/TenantFieldComponent.tsx
@@ -3,9 +3,10 @@
 import type { RelationshipFieldClientProps } from 'payload'
 
 import { RelationshipField, useField, useFormInitializing, useFormModified } from '@payloadcms/ui'
-import { useCallback, useEffect, useRef, useState } from 'react'
+import { useEffect, useRef } from 'react'
 
 import { useTenantSelection } from '@/providers/TenantSelectionProvider/index.client'
+import { useTenantLookup } from '@/utilities/useTenantLookup'
 import './index.scss'
 
 const baseClass = 'tenantField'
@@ -30,58 +31,10 @@ export const TenantFieldComponent = (args: Props) => {
 
   const isGlobalCollection = !!unique
   const hasSetValueRef = useRef(false)
-  // Cache tenant lookups to avoid repeated API calls
-  const [tenantIdBySlug, setTenantIdBySlug] = useState<Record<string, number>>({})
-  const [tenantSlugById, setTenantSlugById] = useState<Record<number, string>>({})
+  const { lookupTenantIdBySlug, lookupTenantSlugById } = useTenantLookup()
 
   // Track whether form has finished initializing
   const formReady = !formInitializing && !formInitializingContext
-
-  const lookupTenantIdBySlug = useCallback(
-    async (slug: string): Promise<number | null> => {
-      if (tenantIdBySlug[slug]) {
-        return tenantIdBySlug[slug]
-      }
-
-      try {
-        const response = await fetch(
-          `/api/tenants?where[slug][equals]=${encodeURIComponent(slug)}&limit=1&depth=0`,
-          { credentials: 'include' },
-        )
-        const result = await response.json()
-        if (result.docs && result.docs.length > 0) {
-          const id = result.docs[0].id
-          setTenantIdBySlug((prev) => ({ ...prev, [slug]: id }))
-          return id
-        }
-      } catch (_) {
-        // Silent fail - tenant field hook will handle missing tenant
-      }
-      return null
-    },
-    [tenantIdBySlug],
-  )
-
-  const lookupTenantSlugById = useCallback(
-    async (id: number): Promise<string | null> => {
-      if (tenantSlugById[id]) {
-        return tenantSlugById[id]
-      }
-
-      try {
-        const response = await fetch(`/api/tenants/${id}?depth=0`, { credentials: 'include' })
-        const result = await response.json()
-        if (result.slug) {
-          setTenantSlugById((prev) => ({ ...prev, [id]: result.slug }))
-          return result.slug
-        }
-      } catch (_) {
-        // Silent fail
-      }
-      return null
-    },
-    [tenantSlugById],
-  )
 
   useEffect(() => {
     // Don't try to set values while the form is still initializing
